the generator

+++ 3dEYE.Benchmark/BenchmarksShared.cs
@@ -5,7 +5,7 @@ namespace _3dEYE.Benchmark;
 
 public static class BenchmarksShared
 {
-    public static async Task GenerateTestFile(this string filePath, ILogger logger, int targetSizeBytes = 1024 * 1024 * 1024)
+    public static async Task GenerateTestFile(this string filePath, ILogger logger, long targetSizeBytes = 1024 * 1024 * 1024)
     {
         var generator = new FileGeneratorFactory(logger).CreateParallelGenerator(filePath, 200 * 1024 * 1024, Environment.ProcessorCount);
         await generator.GenerateFileAsync(filePath, targetSizeBytes);
